double != float
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@8362 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/LocalPlayer.cxx
@@ -1092,7 +1092,7 @@ void			LocalPlayer::jump()
   if (getFlag() == Flags::Wings)
     newVelocity[2] = BZDB.eval(StateDatabase::BZDB_WINGSJUMPVELOCITY);
   else if (getFlag() == Flags::Bouncy) {
-    const float factor = 0.25f + (bzfrand() * 0.75f);
+    const float factor = 0.25f + ((float)bzfrand() * 0.75f);
     newVelocity[2] = factor * BZDB.eval(StateDatabase::BZDB_JUMPVELOCITY);
   }
   else
